Refactoring of most of the redundant code
There's a character class that contains all the logic
for getting the data. A render method to handle the presentation
logic and I tried to split everything by concerns. The ajax
call and iframe loading is at the end.

Hp values are now scraped, so that should fix the current bug.
I worked a little on the table presentation, but I think
something much better could be done by using the DndBeyond's current
css.